kyc_flask/pyproject.toml
[project]
name = "kyc-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-edge-litert>=2.1.3",
    "fastembed>=0.7.4",
    "flask>=3.0",
    "networkx>=3.2",
    "numpy>=1.24",
    "openbharatocr>=0.3.3",
    "opencv-python-headless>=4.10.0",
    "pillow>=10.0",
    "pytesseract>=0.3.10",
    "python-dateutil>=2.9.0.post0",
    "python-levenshtein>=0.23",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.3",
    "scipy>=1.17.1",
    "tensorflow>=2.15",
    "tesseract>=0.1.3",
    "torch>=2.1",
    "werkzeug>=3.0",
]
